+++ cw-dex/src/implementations/pool.rs
@@ -11,9 +11,7 @@ use cosmwasm_std::{Deps, Env, Response, StdResult, Uint128};
 use crate::astroport::AstroportPool;
 
 #[cfg(feature = "osmosis")]
-use {
-    crate::implementations::osmosis::OsmosisPool, apollo_cw_asset::AssetInfoBase, std::str::FromStr,
-};
+use {crate::implementations::osmosis::OsmosisPool, std::str::FromStr};
 
 /// An enum with all known variants that implement the Pool trait.
 /// The ideal solution would of course instead be to use a trait object so that
@@ -51,11 +49,13 @@ impl Pool {
     ///
     /// Arguments:
     /// - `lp_token`: Said LP token
-    pub fn get_pool_for_lp_token(_deps: Deps, lp_token: &AssetInfo) -> Result<Self, CwDexError> {
+    #[allow(unused_variables)]
+    #[allow(unreachable_patterns)]
+    pub fn get_pool_for_lp_token(deps: Deps, lp_token: &AssetInfo) -> Result<Self, CwDexError> {
         match lp_token {
             #[cfg(feature = "osmosis")]
-            AssetInfoBase::Native(lp_token_denom) => {
-                //The only Pool implementation that uses native denoms right now is Osmosis
+            AssetInfo::Native(lp_token_denom) => {
+                // The only Pool implementation that uses native denoms right now is Osmosis
                 if !lp_token_denom.starts_with("gamm/pool/") {
                     return Err(CwDexError::NotLpToken {});
                 }
@@ -66,7 +66,25 @@ impl Pool {
 
                 let pool_id = u64::from_str(pool_id_str).map_err(|_| CwDexError::NotLpToken {})?;
 
-                Ok(Pool::Osmosis(OsmosisPool::new(pool_id, _deps)?))
+                Ok(Pool::Osmosis(OsmosisPool::new(pool_id, deps)?))
+            }
+            #[cfg(feature = "astroport")]
+            AssetInfo::Cw20(address) => {
+                // The only Pool implementation that uses CW20 tokens right now is Astroport.
+                // To figure out if the CW20 is a LP token, we need to check which address
+                // instantiated the CW20 and check if that address is an Astroport pair contract.
+                let contract_info = deps.querier.query_wasm_contract_info(address)?;
+                let creator_addr = deps.api.addr_validate(&contract_info.creator)?;
+
+                // Try to create an `AstroportPool` object with the creator address. This will query
+                // the contract and assume that it is an Astroport pair contract. If it succeeds,
+                // the pool object will be returned.
+                //
+                // NB: This does NOT validate that the pool is registered with the Astroport factory,
+                // and that it is an "official" Astroport pool.
+                let pool = AstroportPool::new(deps, creator_addr)?;
+
+                Ok(Pool::Astroport(pool))
             }
             _ => Err(CwDexError::NotLpToken {}),
         }